Use guacamole for command management
This patch adds dependency on python3-guacamole for working with
commands and sub-commands. Guacamole is in debian and in ubuntu main
archive.

The committer is very fond of dogfooding his own creations and guacamole
can be easily replaced, if desired, as it only plays a part in
organizing command classes that can be nested easily.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>