Add authentication code
This patch adds store authentication module. The module exposes a class
for persistently storing store credentials as well two CLI commands for
logging in and out.

Credentials are stored in $XDG_CONFIG_HOME/ubuntu-image/credentials.ini.
A very simple advisory locking system is used to control concurrent
access should the tool be used as a part of an automated build system.

Credentials are not encrypted. In the future we might adopt a
wallet-like solution that uses the system keyring but this is not
attempted at this stage. At the very least we will have to support fully
unattended installations so this code is useful as-is.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>