Fix our FAT filesystem so that it fits in the partition we're creating
We were creating a 1GiB filesystem image, formatting it, and then only
copying the first 64MiB into the disk image.  This does not yield a working
filesystem! ;)  Only create a 64MiB filesystem, which is what snappy does
currently and is plenty big for our (snappy-specific) purposes.

TODO: fix the code to not hard-code the filesystem sizes in two separate
places.