Fix tests for Xenial docker image.
In Xenial, we need the PyPI flake8-respect-noqa in order to pass the qa
tests, but this is unnecessary in Yakkety and beyond.  And besides, this
package doesn't exist in the archive.

Also, in Xenial's voluptuous, `Coerce` is a function not a class.
Doesn't matter since we really don't need to subclass it anyway.