snapd 2.14.2 compatibility
drop various fields from the test model assertion that are no longer allowed
(or not allowed to be empty).

add a signature to the assertion data - it's an invalid signature taken
from elsewhere, but current snapd no longer allows us to use an assertion
which has no signature at all, giving a parse error.  Even with the SKIP
environment variable, having some signature is now mandatory.