Don't build 4GB images, build images as big as the required contents.
snappy supports auto resizing of the root partition on first boot, so we
don't need to preallocate space.  This makes it faster to write the image to
a disk (we already use sparseness everywhere else to ensure fast writes),
and avoids problems with media that is slightly below the nominal size.

This change:

 - fixes _calculate_dirsize() to always return an integer (rounding up)
 - adds explicit padding for ext4 metadata (so we never try to use a size
   of 0 for our partition when we have no initial files to create)
 - fine-tunes the requirements for the backup GPT space, which previously
   worked more or less by accident
 - does not require the rootfs to be an even multiple of 1MiB, since aligned
   reads don't matter nearly so much at the end of the disk and we'll be
   resized to use whatever space is available anyway so it's short-lived.