Round our rootfs partition up to 1MiB for MBR partitions.
In e660cb5, we relaxed the alignment requirements for the root partition.
However, in addition to an implementation bug that incorrectly caused us to
round down to the nearest 1KiB (instead of up), it appears that sfdisk
forces a 1MiB alignment on partitions and it *also* rounds down.

We ultimately want to fix this to support < 1MiB alignment for partitions,
but in the near term since the underlying tooling is requiring it, fix how
we invoke the partition tools to avoid declaring partitions that are
undersized for our filesystem.

LP: #1634557