assertion_builder: move /boot into system-seed partition for uc20
The current code assumes that /boot can be skipped when the
rootfs is populated. This was a fine assumption for uc16/uc18
but in the uc20 world the only "rootfs" we have is the ubuntu-seed
partition. Here we want to copy the /boot directory with the
grub config so that we have a valid recovery grub installed.